Unified the posts and items services.
The posts-related methods were moved to their own service because
both services got very big; now that most of the methods were
moved away from these services, there's no need for two separate ones.
