+++ src/app/services/items.service.ts
@@ -82,6 +82,10 @@ export class ItemsService {
   idbResolved = {
     user: new BehaviorSubject(false),
   };
+  // Posts variables
+  isUpdated = new BehaviorSubject(false);
+  currentlyOpenMenu = new BehaviorSubject(-1);
+  receivedAHug = new BehaviorSubject(0);
 
   // CTOR
   constructor(
@@ -92,6 +96,131 @@ export class ItemsService {
   ) {}
 
   // POST-RELATED METHODS
+  /*
+  Function Name: sendPost()
+  Function Description: Create a new post.
+  Parameters: post (Post) - the post to attempt to add to the database.
+  ----------------
+  Programmer: Shir Bar Lev.
+  */
+  sendPost(post: Post) {
+    // if the user isn't blocked, let them post
+    if (!this.authService.userData.blocked) {
+      const Url = this.serverUrl + "/posts";
+      this.Http.post(Url, post, {
+        headers: this.authService.authHeader,
+      }).subscribe({
+        next: (response: any) => {
+          this.alertsService.createSuccessAlert(
+            "Your post was published! Return to home page to view the post.",
+            false,
+            "/",
+          );
+          this.alertsService.toggleOfflineAlert();
+
+          let isoDate = new Date(response.posts.date).toISOString();
+          let iDBPost = {
+            ...response.posts,
+            isoDate: isoDate,
+          };
+          this.serviceWorkerM.addItem("posts", iDBPost);
+          // if there was an error, alert the user
+        },
+        error: (err: HttpErrorResponse) => {
+          // if the user is offline, show the offline header message
+          if (!navigator.onLine) {
+            this.alertsService.toggleOfflineAlert();
+          }
+          // otherwise just create an error alert
+          else {
+            this.alertsService.createErrorAlert(err);
+          }
+        },
+      });
+    }
+    // if they're blocked, alert them they cannot post while blocked
+    else {
+      this.alertsService.createAlert({
+        type: "Error",
+        message: `You cannot post new posts while you're blocked. You're blocked until ${this.authService.userData.releaseDate}.`,
+      });
+    }
+  }
+
+  /*
+  Function Name: editPost()
+  Function Description: Edit an existing post. This is used only for editing the post's text.
+  Parameters: post (Post) - the updated data of the post.
+  ----------------
+  Programmer: Shir Bar Lev.
+  */
+  editPost(post: Post) {
+    const Url = this.serverUrl + `/posts/${post.id}`;
+    this.isUpdated.next(false);
+
+    // send update request
+    this.Http.patch(Url, post, {
+      headers: this.authService.authHeader,
+    }).subscribe({
+      next: (_response: any) => {
+        this.alertsService.createSuccessAlert(
+          "Your post was edited. Refresh to view the updated post.",
+          true,
+        );
+        this.alertsService.toggleOfflineAlert();
+        this.isUpdated.next(true);
+        // if there was an error, alert the user
+      },
+      error: (err: HttpErrorResponse) => {
+        // if the user is offline, show the offline header message
+        if (!navigator.onLine) {
+          this.alertsService.toggleOfflineAlert();
+        }
+        // otherwise just create an error alert
+        else {
+          this.alertsService.createErrorAlert(err);
+        }
+      },
+    });
+  }
+
+  /*
+  Function Name: sendHug()
+  Function Description: Send a hug to a user through a post they've written.
+  Parameters: item (Post) - the post for which to send a hug.
+  ----------------
+  Programmer: Shir Bar Lev.
+  */
+  sendHug(item: any) {
+    const Url = this.serverUrl + `/posts/${item.id}/hugs`;
+    this.Http.post(
+      Url,
+      {},
+      {
+        headers: this.authService.authHeader,
+      },
+    ).subscribe({
+      next: (_response: any) => {
+        this.alertsService.createSuccessAlert("Your hug was sent!", false);
+        this.alertsService.toggleOfflineAlert();
+        // Alert the posts that this item received a hug
+        this.receivedAHug.next(item.id);
+        // if there was an error, alert the user
+      },
+      error: (err: HttpErrorResponse) => {
+        // if the user is offline, show the offline header message
+        if (!navigator.onLine) {
+          this.alertsService.toggleOfflineAlert();
+        }
+        // otherwise just create an error alert
+        else {
+          this.alertsService.createErrorAlert(err);
+        }
+      },
+    });
+  }
+
+  // USER-RELATED METHODS
   // ==============================================================
   /*
   Function Name: sendUserHug()
@@ -131,8 +260,6 @@ export class ItemsService {
     });
   }
 
-  // USER-RELATED METHODS
-  // ==============================================================
   /*
   Function Name: getUser()
   Function Description: Gets the data of a specific user.
